IS-2195: Fix alert bug
